Add 'server.rs'
